--- practice2025/task10/task10.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.10.35122.118
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "task10", "task10\task10.csproj", "{C7C5398E-DF9B-440F-8CB1-3D67FC25877C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C7C5398E-DF9B-440F-8CB1-3D67FC25877C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C7C5398E-DF9B-440F-8CB1-3D67FC25877C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C7C5398E-DF9B-440F-8CB1-3D67FC25877C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C7C5398E-DF9B-440F-8CB1-3D67FC25877C}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {9C0E9AB5-3A30-49B3-B673-4B19060E9A8E}
-	EndGlobalSection
-EndGlobal
